WIP: Implement a method of retrieving the package update time
The package update time is based on the last update of the _source_
pacakge, not the binary package.  To do this, I have a server
application, currently hosted on piggz.co.uk:8081 which uses the OBS API
to build a JSON model, relating projects, packages, repositories and
binaries.

To find the package mtime, you iterate the model, searching for the
project, then find the repository matching the current device, then look
for the package in the binaries.  WHen the pacakge is found from the
binary, get the mtime from the source package list.

Currently I dont know how to get the repo name from SSU, therfore the
arch is hardcoded as aarch64.

Included some simple update to the QML to show the update time, nothing
more complicated such as a section of "recently udpated apps"

Server code will be posted later for comment, it could certainly be
improved (currently a single large cache), and Im happy for others to
take this on if the idea is sound.